Settings: Refactor module namespace to make room for sibling module
