+++ tests/docs/test_settings.py
@@ -91,45 +91,3 @@ def test_settings_yaml(tmp_path: Path):
     # Verify the outcome.
     data = yaml.safe_load(output_path.read_text())
     assert "whether or not to collect statistical information" in data["settings"]["stats.enabled"]["purpose"]
-
-
-def test_functions_json(tmp_path: Path):
-    """
-    Verify `ctk docs functions`.
-    """
-
-    output_path = tmp_path / "cratedb-functions.json"
-
-    # Invoke command.
-    runner = CliRunner()
-    result = runner.invoke(
-        cli,
-        args=f"functions --format=json --output={output_path}",
-        catch_exceptions=False,
-    )
-    assert result.exit_code == 0
-
-    # Verify the outcome.
-    data = json.loads(output_path.read_text())
-    assert "substr('string' FROM 'pattern')" in data["functions"]
-
-
-def test_functions_markdown(tmp_path: Path):
-    """
-    Verify `ctk docs functions`.
-    """
-
-    output_path = tmp_path / "cratedb-functions.md"
-
-    # Invoke command.
-    runner = CliRunner()
-    result = runner.invoke(
-        cli,
-        args=f"functions --format=markdown --output={output_path}",
-        catch_exceptions=False,
-    )
-    assert result.exit_code == 0
-
-    # Verify the outcome.
-    data = output_path.read_text()
-    assert "substr('string' FROM 'pattern')" in data
